backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bottle==0.12.25
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
exa-py==1.0.9
fastapi==0.110.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Mako==1.3.3
MarkupSafe==2.1.5
openai==1.25.0
pongo-python==3.0.0
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.1
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
websockets==12.0